requirements.txt
flask
flask-sqlalchemy
flask-assets
flask-script
flask-redis
flask-migrate
cssmin
jsmin
raven
blinker
oauth2client